+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './ButtonSelect';
 export * from './editors';
 export * from './Table';
 export * from './TablePanel';